+++ src/AtomUI.Desktop.Controls/CheckBox/CheckBox.SemanticParts.cs
@@ -1,0 +1,19 @@
+using AtomUI.Theme;
+using Avalonia.Controls.Presenters;
+using Avalonia.Controls.Primitives;
+
+namespace AtomUI.Desktop.Controls;
+
+[SemanticPart(
+    "icon",
+    SelectorClass = "semantic-icon",
+    ContractType = typeof(TemplatedControl),
+    Since = "6.0")]
+[SemanticPart(
+    "label",
+    SelectorClass = "semantic-label",
+    ContractType = typeof(ContentPresenter),
+    Since = "6.0")]
+public partial class CheckBox
+{
+}