+++ src/AtomUI.Desktop.Controls/CheckBox/CheckBox.cs
@@ -2,7 +2,7 @@ using AtomUI.Controls.Commons;
 
 namespace AtomUI.Desktop.Controls;
 
-public class CheckBox : AbstractCheckBox
+public partial class CheckBox : AbstractCheckBox
 {
     public CheckBox()
     {